+++ src/test/main.ts
@@ -15,6 +15,7 @@ import "./input.test.js";
 import "./regression.test.js";
 import "./integration.test.js";
 import "./ui.test.js";
+import "./round5.test.js";
 
 import { run } from "./runner.js";
 run();